[Issue #1524] Cleanup the opportunity route tests
