+++ api/tests/src/api/opportunities_v0_1/test_opportunity_route_get.py
@@ -1,0 +1,118 @@
+import pytest
+
+from src.db.models.opportunity_models import Opportunity
+from tests.src.api.opportunities_v0_1.conftest import get_search_request, validate_opportunity
+from tests.src.db.models.factories import (
+    CurrentOpportunitySummaryFactory,
+    OpportunityFactory,
+    OpportunitySummaryFactory,
+)
+
+
+@pytest.fixture
+def truncate_opportunities(db_session):
+    # Note that we can't just do db_session.query(Opportunity).delete() as the cascade deletes won't work automatically:
+    # https://docs.sqlalchemy.org/en/20/orm/queryguide/dml.html#orm-queryguide-update-delete-caveats
+    # but if we do it individually they will
+    opportunities = db_session.query(Opportunity).all()
+    for opp in opportunities:
+        db_session.delete(opp)
+
+    # Force the deletes to the DB
+    db_session.commit()
+
+
+#####################################
+# GET opportunity tests
+#####################################
+
+
+@pytest.mark.parametrize(
+    "opportunity_params,opportunity_summary_params",
+    [
+        ({}, {}),
+        # Only an opportunity exists, no other connected records
+        (
+            {
+                "opportunity_assistance_listings": [],
+            },
+            None,
+        ),
+        # Summary exists, but none of the list values set
+        (
+            {},
+            {
+                "link_funding_instruments": [],
+                "link_funding_categories": [],
+                "link_applicant_types": [],
+            },
+        ),
+        # All possible values set to null/empty
+        # Note this uses traits on the factories to handle setting everything
+        ({"all_fields_null": True}, {"all_fields_null": True}),
+    ],
+)
+def test_get_opportunity_200(
+    client, api_auth_token, enable_factory_create, opportunity_params, opportunity_summary_params
+):
+    # Split the setup of the opportunity from the opportunity summary to simplify the factory usage a bit
+    db_opportunity = OpportunityFactory.create(
+        **opportunity_params, current_opportunity_summary=None
+    )  # We'll set the current opportunity below
+
+    if opportunity_summary_params is not None:
+        db_opportunity_summary = OpportunitySummaryFactory.create(
+            **opportunity_summary_params, opportunity=db_opportunity
+        )
+        CurrentOpportunitySummaryFactory.create(
+            opportunity=db_opportunity, opportunity_summary=db_opportunity_summary
+        )
+
+    resp = client.get(
+        f"/v0.1/opportunities/{db_opportunity.opportunity_id}", headers={"X-Auth": api_auth_token}
+    )
+    assert resp.status_code == 200
+    response_data = resp.get_json()["data"]
+
+    validate_opportunity(db_opportunity, response_data)
+
+
+def test_get_opportunity_404_not_found(client, api_auth_token, truncate_opportunities):
+    resp = client.get("/v0.1/opportunities/1", headers={"X-Auth": api_auth_token})
+    assert resp.status_code == 404
+    assert resp.get_json()["message"] == "Could not find Opportunity with ID 1"
+
+
+def test_get_opportunity_404_not_found_is_draft(client, api_auth_token, enable_factory_create):
+    # The endpoint won't return drafts, so this'll be a 404 despite existing
+    opportunity = OpportunityFactory.create(is_draft=True)
+
+    resp = client.get(
+        f"/v0.1/opportunities/{opportunity.opportunity_id}", headers={"X-Auth": api_auth_token}
+    )
+    assert resp.status_code == 404
+    assert (
+        resp.get_json()["message"]
+        == f"Could not find Opportunity with ID {opportunity.opportunity_id}"
+    )
+
+
+#####################################
+# Auth tests
+#####################################
+@pytest.mark.parametrize(
+    "method,url,body",
+    [
+        ("POST", "/v0.1/opportunities/search", get_search_request()),
+        ("GET", "/v0.1/opportunities/1", None),
+    ],
+)
+def test_opportunity_unauthorized_401(client, api_auth_token, method, url, body):
+    # open is just the generic method that post/get/etc. call under the hood
+    response = client.open(url, method=method, json=body, headers={"X-Auth": "incorrect token"})
+
+    assert response.status_code == 401
+    assert (
+        response.get_json()["message"]
+        == "The server could not verify that you are authorized to access the URL requested"
+    )
